refactor: rename protocols.py to _typing.py
This makes sense as the file does not contain protocols, just things
related to typing of the code.
