+++ src/shapleypy/solution_concept/shapley_value.py
@@ -6,9 +6,9 @@ from typing import Any
 
 import numpy as np
 
+from shapleypy._typing import Player, Value
 from shapleypy.coalition import Coalition
 from shapleypy.game import Game
-from shapleypy.protocols import Player, Value
 from shapleypy.solution_concept._default_value import set_default_value
